+++ apps/react-namespace/src/components/providerHelper/WithComposeProvider.tsx
@@ -3,14 +3,14 @@ import React, { ComponentType, FC, ReactElement } from 'react'
 import ComposeProviders from './ComposeProviders'
 
 export type WithComposedProviderProps<C> = {
-  Component: ComponentType<C>
-  providers: ReactElement[]
+  Components: ComponentType<C>
+  Providers: ReactElement[]
 }
 
-const WithComposedProviders = <C extends object>({ Component, providers }: WithComposedProviderProps<C>): FC<C> => {
+const WithComposedProviders = <C extends object>({ Components, Providers }: WithComposedProviderProps<C>): FC<C> => {
   const WrappedComponent: FC<C> = (props: C) => (
-    <ComposeProviders providers={providers}>
-      <Component {...props} />
+    <ComposeProviders providers={Providers}>
+      <Components {...props} />
     </ComposeProviders>
   )
   return WrappedComponent