+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @lodado/react-namespace-storybook-docs
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @lodado/react-namespace@0.1.6
+
 ## 0.0.6
 
 ### Patch Changes